Move the factory to an abstract base class
