+++ src/Composer/Action/ShowPackage.php
@@ -18,25 +18,10 @@ use Silex\Application;
  *
  * @author Gawain Lynch <gawain.lynch@gmail.com>
  */
-final class ShowPackage
+final class ShowPackage extends BaseAction
 {
-    /**
-     * @var \Silex\Application
-     */
-    private $app;
-
-    /**
-     * @var \Composer\Package\Version\VersionParser
-     */
+    /** @var \Composer\Package\Version\VersionParser */
     private $versionParser;
-
-    /**
-     * @param $app \Silex\Application
-     */
-    public function __construct(Application $app)
-    {
-        $this->app = $app;
-    }
 
     /**
      * Retrieves detailed information about a package, or lists all packages available.
