+++ src/Provider/ExtensionServiceProvider.php
@@ -3,7 +3,6 @@
 namespace Bolt\Provider;
 
 use Bolt\Composer\Action;
-use Bolt\Composer\Factory;
 use Bolt\Composer\PackageManager;
 use Bolt\Extensions;
 use Bolt\Extensions\ExtensionsInfoService;
@@ -48,11 +47,6 @@ class ExtensionServiceProvider implements ServiceProviderInterface
                 return new PackageManager($app);
             }
         );
-        $app['extend.factory'] = $app->share(
-            function ($app) {
-                return new Factory($app['extend.manager']->getOptions(), $app['logger.system']);
-            }
-        );
 
         $app['extend.info'] = $app->share(
             function ($app) {
@@ -67,7 +61,7 @@ class ExtensionServiceProvider implements ServiceProviderInterface
                 'autoload' => $app->share(function () use ($app) { return new Action\DumpAutoload($app); }),
                 'check'    => $app->share(function () use ($app) { return new Action\CheckPackage($app); }),
                 'install'  => $app->share(function () use ($app) { return new Action\InstallPackage($app); }),
-                'json'     => $app->share(function () use ($app) { return new Action\BoltExtendJson($app['extend.manager']->getOptions()); }),
+                'json'     => $app->share(function () use ($app) { return new Action\BoltExtendJson($app); }),
                 'remove'   => $app->share(function () use ($app) { return new Action\RemovePackage($app); }),
                 'require'  => $app->share(function () use ($app) { return new Action\RequirePackage($app); }),
                 'search'   => $app->share(function () use ($app) { return new Action\SearchPackage($app); }),
